did some exercise

+++ db/seeds.rb
@@ -1,16 +1,18 @@
-puts 'Cleaning database...'
-Cocktail.destroy_all
+require 'open-uri'
 
-puts 'Creating cocktails...'
-Cocktail.create(name: "Rum soda")
-Cocktail.create(name: "Pinacorada")
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
 
-Ingredient.create(name: "lemon")
-Ingredient.create(name: "ice")
-Ingredient.create(name: "mint leaves")
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
 
-Cocktail.create!(@cocktail)
-puts 'Finished!'
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
+end
